add color chooser

+++ Build.sh
@@ -2,6 +2,7 @@
 
 # @author Henrik Lundgren
 # @date 2014-01-23
+# @url github.com/henriklundgren/deck
 
 # Variables
 temp=tmp
@@ -15,16 +16,45 @@ titles=('title-3-active' 'title-3-inactive')
 buttons=('active' 'pressed')
 signs=('close' 'hide' 'maximize' 'menu' 'shade' 'stick')
 
-echo 'Welcome to Deck theme creator'
-echo 'I require xcftools to process GIMP xcf file format and Imagemagick to process the rest of the steps.'
-
-read -p 'Press [ENTER] to continue...'
 clear
 
-xcfinfo $src
+intro='
+--------------------------------------------
+Welcome to Deck theme creator.
+============================================
+\E[32m          __________________________
+ ________|   ____           _       |_______
+ \       |  (|   \Xfwm     | |      |      /
+  \      |   |    | _   __ | |      |     /
+   \     |  _|    ||/  /   |/_)     |    /
+   /     | (/\___/ |__/\__/| \_/\E[31mv0.1\E[32m|    \
+  /      |__________________________|     \
+ /__________)                    (_________\
 
-read -p 'Press [ENTER] to continue...'
-clear
+\E[0m
+============================================
+Script to build and deploy Deck Xfwm theme
+on any Xfce GNU/Linux OS.
+
+More information about Deck and this script
+can be found on the Github page.
+http://github.com/henriklundgren/deck
+
+I will extract layers from xcf
+source file and prepare sprite
+for distribution.
+-------------------------------------------
+'
+echo -e "$intro"
+
+# Make sure requirements are meet.
+command -v xcfinfo >/dev/null 2>&1 || { echo "Require Xcftools but it's not installed. Aborting." >&2; exit 1; }
+command -v convert >/dev/null 2>&1 || { echo "Require Imagemagick but it's not installed. Aborting." >&2; exit 1; }
+
+read -p "Press [ENTER] to continue..."
+
+# Put output in variable and make test to see if source has all layers
+#xcfinfo -v $src
 
 # remove old (should first check if they exist I guess)
 rm -r $temp
@@ -34,36 +64,44 @@ rm -r $dist
 mkdir $temp
 mkdir -p $dist
 
+echo -e "
+======================================
+The default frame has a lightgrey base
+color. You can change this now if you
+like to.
+--------------------------------------
+"
+echo -en "What color do you want [lightgrey]: "; read ask_color
 
-# Base frame active
+# Base frame
 ########################
 
-for frame in ${framesActive[@]}
-do
+for frame in ${framesActive[@]}; do
   # Extract xpm
-  xcf2png -vC $src "$frame".xpm -o $temp/"$frame".xpm.temp
+  xcf2png -C   $src  "$frame".xpm    -o $temp/"$frame".xpm.temp
+  # Convert PNG to XPM
+  convert $temp/"$frame".xpm.temp   +map    $temp/"$frame".xpm
+
+  # Change base color
+  if [ "$ask_color" ]; then
+    convert  $temp/"$frame".xpm -alpha extract -background "$ask_color" -alpha shape $temp/"$frame".xpm
+  fi
+done
+
+for frame in ${framesInactive[@]}; do
+  # Extract xpm
+  xcf2png -C $src "$frame".xpm -o $temp/"$frame".xpm.temp
 
   # Convert PNG to XPM
   convert $temp/"$frame".xpm.temp +map $temp/"$frame".xpm
 
-  # Remove unused for now only (later we only remove the entire folder)
-  #rm $temp/"$frame".xpm.temp
+  # Change base color
+  if [ "$ask_color" ]; then
+    convert  $temp/"$frame".xpm -alpha extract -background "$ask_color" -alpha shape $temp/"$frame".xpm
+  fi
 done
 
-for frame in ${framesInactive[@]}
-do
-  # Extract xpm
-  xcf2png -vC $src "$frame".xpm -o $temp/"$frame".xpm.temp
-
-  # Convert PNG to XPM
-  convert $temp/"$frame".xpm.temp +map $temp/"$frame".xpm
-
-  # Remove unused for now only (later we only remove the entire folder)
-  #rm $temp/"$frame".xpm.temp
-done
-
-for val in ${framesActive[@]}
-do
+for val in ${framesActive[@]}; do
   # Skip bottom
   if [ "$val" = "bottom-active" ]
   then
@@ -79,25 +117,106 @@ do
   fi
 done
 
-for vall in ${framesInactive[@]}
+for val in ${framesInactive[@]}
 do
   # Skip bottom
-  if [ "$vall" = "bottom-inactive" ]
+  if [ "$val" = "bottom-inactive" ]
   then
     continue
-  elif [ "$vall" = "title-2-inactive" ]
+  elif [ "$val" = "title-2-inactive" ]
   then
-    convert $temp/"$vall".xpm -flop $temp/${vall/2/4}.xpm
-  elif [ "$vall" = "title-1-inactive" ]
+    convert $temp/"$val".xpm -flop $temp/${val/2/4}.xpm
+  elif [ "$val" = "title-1-inactive" ]
   then
-    convert $temp/"$vall".xpm -flop $temp/${vall/1/5}.xpm
+    convert $temp/"$val".xpm -flop $temp/${val/1/5}.xpm
   else
-    convert $temp/"$vall".xpm -flop $temp/${vall/left/right}.xpm
+    convert $temp/"$val".xpm -flop $temp/${val/left/right}.xpm
   fi
 done
 
-ls --ignore=*.temp $temp
-read -p 'Base frame is done. Press [RETURN] to continue...'
+
+
+xpm_file_array=($(ls --ignore=*.temp $temp)) # http://stackoverflow.com/a/13824135
+xpm_base_array=([1]=bottom [2]=bottom-left [3]=left [4]=top-left [5]=top-right
+[6]=right [7]=bottom-right [8]=title-1 [9]=title-2 [10]=title-4 [11]=title-5);
+
+#echo -en "The number of files found: ${#xpm_file_array[*]}"
+
+# Function to check output files with expected result
+check_if_xpm_exist() {
+  local switch=false
+  for item in ${xpm_file_array[@]}; do
+    if [ "$item" = "$1"-"$2".xpm ]; then
+      switch=true
+      echo -en "[\E[0;32mOK\E[0m]"
+    fi
+  done
+
+  if [ "$switch" == false ]; then
+    echo -en "[\E[0;31m--\E[0m]"
+  fi
+}
+
+printf '\n%-30s\n' "-------------------------------------"
+printf '\e[37;40m%-30s\e[0m\n' "|:              Output             :|"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| " "| Active " "| Inactive |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[1]}" "|  $(check_if_xpm_exist ${xpm_base_array[1]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[1]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[2]}" "|  $(check_if_xpm_exist ${xpm_base_array[2]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[2]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[3]}" "|  $(check_if_xpm_exist ${xpm_base_array[3]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[3]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[4]}" "|  $(check_if_xpm_exist ${xpm_base_array[4]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[4]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[5]}" "|  $(check_if_xpm_exist ${xpm_base_array[5]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[5]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[6]}" "|  $(check_if_xpm_exist ${xpm_base_array[6]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[6]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[7]}" "|  $(check_if_xpm_exist ${xpm_base_array[7]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[7]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[8]}" "|  $(check_if_xpm_exist ${xpm_base_array[8]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[8]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[9]}" "|  $(check_if_xpm_exist ${xpm_base_array[9]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[9]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[10]}" "|  $(check_if_xpm_exist ${xpm_base_array[10]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[10]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+printf '%-14s %-8s %-8s\n' "| ${xpm_base_array[11]}" "|  $(check_if_xpm_exist ${xpm_base_array[11]} active)  " "|   $(check_if_xpm_exist ${xpm_base_array[11]} inactive)   |"
+printf '%-30s\n' "-------------------------------------"
+
+
+
+
+#base_files=`ls --ignore=*.temp $temp`
+#echo -n "$base_files"
+
+if [ "${#xpm_file_array[@]}" = 22 ]
+then
+  echo -e "
+*************************************
+\e[31m#####################################
+#### ### ###     #####  #####  ######
+##### # ####      ####   ###   ######
+###### #####     ##### #  #  # ######
+##### # #### ######### ##   ## ######
+###  ### ### ######### ####### ######
+#####################################\e[0m
+=====================================
+Correct amount of sprites found.
+  "
+  echo -en ${#xpm_file_array[@]}
+else
+  echo "
+===================================
+Something went wrong. Aborting.
+  "
+  exit 1
+fi
+
+read -p "
+Press [RETURN] to continue..."
+clear
 
 
 
@@ -107,13 +226,13 @@ read -p 'Base frame is done. Press [RETURN] to continue...'
 for frame in ${framesActive[@]}
 do
   # Extract
-  xcf2png -vCA $src "$frame".png -o $temp/"$frame".png.temp
+  xcf2png -CA $src "$frame".png -o $temp/"$frame".png.temp
 done
 
 for frame in ${framesInactive[@]}
 do
   # Extract
-  xcf2png -vCA $src "$frame".png -o $temp/"$frame".png.temp
+  xcf2png -CA $src "$frame".png -o $temp/"$frame".png.temp
 done
 
 for val in ${framesActive[@]}
@@ -154,9 +273,10 @@ do
   convert $temp/"$val".png.temp -channel Alpha -evaluate Divide 2 png32:$temp/"$val".png
 done
 
-ls --ignore=*.temp $temp
+ls --ignore='*.temp' --ignore='*.xpm' $temp
 read -p 'Overlay frame is done. Press [RETURN] to continue...'
 clear
+
 
 # Title plate
 #################
